REFACTOR pull requested changes
